Fix | Fixed error from repository and service, entity and dto folder created in model folder
