+++ backend/src/main/java/com/moshop/backend/model/entity/Product.java
@@ -1,4 +1,4 @@
-package com.moshop.backend.model;
+package com.moshop.backend.model.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -6,20 +6,21 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
-
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 
-@Document
 @Data
-@Builder
-@NoArgsConstructor
 @AllArgsConstructor
-public class Cart {
+@NoArgsConstructor
+@Builder
+@Document
+public class Product {
     @Id
-    private String cartId;
-    private String customerId;
     private String productId;
+    private String productName;
+    private double productPrice;
+    private String productDescription;
+    private String productImage;
+    private String variantId;
     private LocalDateTime createdDate;
     private LocalDateTime updatedDate;
     private boolean isActive;
