+++ backend/src/main/java/com/moshop/backend/repository/CartRepository.java
@@ -1,11 +1,9 @@
 package com.moshop.backend.repository;
 
-import com.moshop.backend.model.Cart;
+import com.moshop.backend.model.entity.Cart;
 import org.springframework.data.mongodb.repository.MongoRepository;
-import org.springframework.stereotype.Repository;
 
 import java.util.List;
-import java.util.Optional;
 
 public interface CartRepository extends MongoRepository<Cart, String> {
     List<Cart> findAllByCustomerId(String customerId);
