+++ backend/src/main/java/com/moshop/backend/repository/CustomerRepository.java
@@ -1,13 +1,12 @@
 package com.moshop.backend.repository;
 
-import com.moshop.backend.model.Customer;
+import com.moshop.backend.model.entity.Customer;
 import org.springframework.data.mongodb.repository.MongoRepository;
-import org.springframework.data.mongodb.repository.Query;
 
 import java.util.List;
 import java.util.Optional;
 
 public interface CustomerRepository extends MongoRepository <Customer, String> {
-    List<Customer> findByIsActiveTrue();
+    List<Customer> findAllByIsActive(boolean isActive);
     Optional<Customer> findByCustomerEmailAndCustomerPassword(String email, String password);
 }
