+++ backend/src/main/java/com/moshop/backend/repository/ProductRepository.java
@@ -1,6 +1,6 @@
 package com.moshop.backend.repository;
 
-import com.moshop.backend.model.Product;
+import com.moshop.backend.model.entity.Product;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
@@ -8,7 +8,5 @@ import java.util.List;
 
 @Repository
 public interface ProductRepository extends MongoRepository<Product, String> {
-    List<Product> findAllByIsActiveProduct(boolean isActive);
-
-    void delete(boolean isActive);
+    List<Product> findAllByIsActive(boolean isActive);
 }
