+++ backend/src/main/java/com/moshop/backend/repository/VariantRepository.java
@@ -1,6 +1,6 @@
 package com.moshop.backend.repository;
 
-import com.moshop.backend.model.Variant;
+import com.moshop.backend.model.entity.Variant;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
@@ -8,7 +8,5 @@ import java.util.List;
 
 @Repository
 public interface VariantRepository extends MongoRepository<Variant, String> {
-    List<Variant> findByAllIsActive(boolean isActive);
-
-    void delete(boolean isActive);
+    List<Variant> findAllByIsActive(boolean isActive);
 }
