+++ backend/src/main/java/com/moshop/backend/services/OrderService.java
@@ -1,6 +1,6 @@
 package com.moshop.backend.services;
 
-import com.moshop.backend.model.Order;
+import com.moshop.backend.model.entity.Order;
 
 import java.util.List;
